[cluster_test] Use debug port to fetch structured logs
Kinesis has a very wide distribution of log delays, to reliably run cluster_test with it we would have to set our liveness check boundaries to 2 minutes, to simply account for kinesis reliability
This is not really acceptable, so this PR uses debug port to fetch logs directly from validator, bypassing Cloud Watch and Kinesis.
This allows to reduce liveness check window to 15 seconds (potentially even lower, but 15 seconds is good enough for our use case). Liveness window is now driven by design and timeouts in our system, rather then latency of log pipeline

This PR also adds `--peers` argument to cluster_test and `Cluster::sub_cluster` method, to run cluster test on subset of nodes (helpful for debugging/demonstration)
